Reorganize file directory structure
