+++ apps/docs/libraries/ts/redwoodjs/keys.mdx
@@ -1,0 +1,8 @@
+---
+title: "API Key Verification"
+description: "Implement API Key Verification with RedwoodJS Middleware and Unkey"
+---
+
+# Usage
+
+Coming soon.
