+++ apps/docs/libraries/ts/redwoodjs/overview.mdx
@@ -7,12 +7,22 @@ The official RedwoodJS SDK for Unkey.
 
 - Rate limit requests using `createRatelimitMiddleware` middleware
 
+## Rate Limiting
+
+RedwoodJS supports Unkey rate limiting via:
+
+- middleware with Server-Side Rendering (SSR) enabled
+
+In the future, support for rate limiting GraphQL operations and other RedwoodJS functions will be added.
+
+## API Key Verification
+
+RedwoodJS support for Unkey API key verification is coming soon.
+
 ## Install
 
 <Tabs>
   <Tab title="yarn">
-    ```
-    bash yarn rw workspace web add @unkey/redwoodjs
-    ```
+    ``` bash yarn rw workspace web add @unkey/redwoodjs ```
   </Tab>
 </Tabs>
