+++ packages/redwoodjs/src/index.ts
@@ -1,5 +1,11 @@
 import type { Logger } from "@redwoodjs/api/logger";
 
-export * from "./ratelimit";
+export * from "./keys/middleware/createApiKeyMiddleware";
+
+export * from "./ratelimit/middleware/createRatelimitMiddleware";
+export * from "./ratelimit/middleware/util";
+
+export type * from "./keys/middleware/types";
+export type * from "./ratelimit/middleware/types";
 
 export const defaultLogger = require("abstract-logging") as Logger;
