+++ packages/redwoodjs/src/keys/middleware/createApiKeyMiddleware.ts
@@ -1,7 +1,7 @@
 import { Unkey } from "@unkey/api";
-import { version } from "../../package.json";
+import { version } from "../../../package.json";
 
-import { defaultLogger } from "../index";
+import { defaultLogger } from "../../index";
 
 import type { Middleware, MiddlewareRequest, MiddlewareResponse } from "@redwoodjs/vite/middleware";
 import type { ApiKeyMiddlewareConfig } from "./types";
